Add gates to all controller endpoints

+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ProductCreateRequest;
 use App\Http\Resources\ProductResource;
 use App\Product;
+use Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Symfony\Component\HttpFoundation\Response;
@@ -13,6 +14,8 @@ class ProductController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view', 'products');
+
         $products = Product::paginate();
 
         return ProductResource::collection($products);
@@ -20,11 +23,15 @@ class ProductController extends Controller
 
     public function show($id)
     {
+        Gate::authorize('view', 'products');
+
         return new ProductResource(Product::find($id));
     }
 
     public function store(ProductCreateRequest $request)
     {
+        Gate::authorize('edit', 'products');
+
         $product = Product::create($request->only('title', 'description', 'price', 'image'));
 
         return response($product, Response::HTTP_CREATED);
@@ -32,6 +39,8 @@ class ProductController extends Controller
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'products');
+
         $product = Product::find($id);
 
         $product->update($request->only('title', 'description', 'price', 'image'));
@@ -41,6 +50,8 @@ class ProductController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('edit', 'products');
+
         Product::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
